adjust diretory
