+++ include/aquarius_protocol.hpp
@@ -1,9 +1,8 @@
 #pragma once
-#include <aquarius_protocol/basic_message.hpp>
 #include <aquarius_protocol/binary.hpp>
 #include <aquarius_protocol/concepts.hpp>
-#include <aquarius_protocol/ip/tcp.hpp>
-#include <aquarius_protocol/ip/tcp/header.hpp>
 #include <aquarius_protocol/json.hpp>
 #include <aquarius_protocol/reflect_helper.hpp>
 #include <aquarius_protocol/reflection.hpp>
+#include <aquarius_protocol/tcp_request.hpp>
+#include <aquarius_protocol/tcp_response.hpp>
